+++ functional-tests/plege_sector_test.go
@@ -42,7 +42,7 @@ func TestMiningPledgeSector(t *testing.T) {
 	chainClock := clock.NewChainClockFromClock(uint64(genTime), blockTime, fakeClock)
 
 	drandImpl := &drand.Fake{
-		GenesisTime:   time.Unix(int64(genTime), 0).Add(-1 * blockTime),
+		GenesisTime:   time.Unix(genTime, 0).Add(-1 * blockTime),
 		FirstFilecoin: 0,
 	}
